fix: 자료형 변경 (Integer -> Long)

+++ src/main/java/click/dailyfeed/search/comments/projection/PostCommentCountProjection.java
@@ -11,5 +11,5 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class PostCommentCountProjection {
     private Long postPk;
-    private Integer commentCount;
+    private Long commentCount;
 }
